+++ extension/lib/queries/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  getCustomer: require('./getCustomer')
+}