+++ extension/models/user/ShopgateCustomer.js
@@ -44,7 +44,12 @@ class ShopgateCustomer {
       shopifyCustomer.firstName,
       shopifyCustomer.lastName,
       {
-        phone: shopifyCustomer.phoneNumber
+        phone: shopifyCustomer.phoneNumber,
+        shopifyCompanyContacts: ((shopifyCustomer.companyContacts || {}).edges || []).map(companyContact => ({
+            id: companyContact.node.id,
+            name: companyContact.node.company.name,
+            locations: ((companyContact.node.locations || {}).edges || []).map(location => location.node)
+        }))
       }
     )
   }